+++ sdwan_policy_objects.tf
@@ -153,6 +153,15 @@ resource "sdwan_policy_object_security_data_ipv4_prefix_list" "policy_object_sec
   }]
 }
 
+resource "sdwan_policy_object_security_fqdn_list" "policy_object_security_fqdn_list" {
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.security_fqdn_lists, {}) : p.name => p }
+  name               = each.value.name
+  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  entries = [for e in try(each.value.fqdns, []) : {
+    pattern = try(e, null)
+  }]
+}
+
 resource "sdwan_policy_object_sla_class_list" "policy_object_sla_class_list" {
   for_each           = { for p in try(local.feature_profiles.policy_object_profile.sla_classes, {}) : p.name => p }
   name               = each.value.name